requirements.txt
absl-py==2.1.0
accelerate==1.2.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
args==0.1.0
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
blis==1.2.0
braceexpand==0.1.7
cachetools==5.5.2
catalogue==2.0.10
certifi==2024.12.14
cfgv==3.4.0
charset-normalizer==3.4.1
clean-fid==0.1.35
click==8.1.8
clint==0.5.1
clip==0.2.0
cloudpathlib==0.21.0
comm==0.2.2
confection==0.1.5
contexttimer==0.3.3
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
datasets==3.5.0
debugpy==1.8.11
decorator==5.1.1
decord==0.6.0
diffusers==0.32.2
dill==0.3.8
distlib==0.3.9
distro==1.9.0
docker-pycreds==0.4.0
einops==0.8.1
exceptiongroup==1.2.2
executing==2.1.0
fairscale==0.4.13
faiss-cpu==1.10.0
faiss-gpu==1.7.2
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.16.1
fire==0.4.0
flatbuffers==24.3.25
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
gradio==5.23.1
gradio_client==1.8.0
groovy==0.1.2
grpcio==1.68.0
h11==0.14.0
h5py==3.12.1
hpsv2==1.2.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
identify==2.6.9
idna==3.10
image-reward==1.5
imageio==2.37.0
importlib_metadata==8.5.0
iniconfig==2.1.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.15
kaggle==1.7.4.2
keras==3.6.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
lightning==2.5.0.post0
lightning-utilities==0.11.9
marisa-trie==1.2.1
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
namex==0.0.8
narwhals==1.32.0
nest_asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
open_clip_torch==2.32.0
openai==1.68.2
opencv-python==4.11.0.86
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.16
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.15.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pip==25.0.1
platformdirs==4.3.6
plotly==6.0.1
pluggy==1.5.0
portalocker==3.1.1
pre_commit==4.2.0
preshed==3.0.9
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==3.20.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pycocoevalcap==1.2
pycocotools==2.0.8
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.1
PySocks==1.7.1
pytest==7.2.0
pytest-split==0.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.20
python-slugify==8.0.4
pytorch-lightning==2.5.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
ruff==0.11.2
safehttpx==0.1.6
safetensors==0.4.5
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.0
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.20.0
setproctitle==1.3.4
setuptools==75.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack_data==0.6.3
starlette==0.46.1
streamlit==1.44.0
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
text-unidecode==1.3
thinc==8.3.4
threadpoolctl==3.6.0
tifffile==2025.3.13
tiktoken==0.9.0
timm==0.6.13
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
torch==2.5.1
torch-fidelity==0.3.0
torchmetrics==1.6.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
triton==3.1.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.3
wandb==0.19.4
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webdataset==0.2.111
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.44.0
widgetsnbextension==4.0.14
wrapt==1.17.0
xformers==0.0.29.post1
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
